services/inventory/src/services/web/requirements.txt
Django==2.0.7
pymongo==3.7.1
python-dateutil==2.6.0
pytz==2018.5
six==1.11.0
uWSGI==2.0.17.1
